:recycle: Reorganize compat support files under spec/support/compat/
- Move fluent_compat_base.rb to compat/base.rb
- Move fluent_syntax/ to compat/syntax/
- Move fluent_bundle/ to compat/bundle/
- Move fixture configuration from test_helper to spec files
- Add collect_fixtures method to unify fixture loading
