+++ spec/compat/syntax_parser_spec.rb
@@ -1,11 +1,20 @@
 # frozen_string_literal: true
 
-require_relative "../support/fluent_syntax"
+require_relative "../support/compat/syntax"
+
+FIXTURES_ROOT = FluentCompatBase::FLUENT_JS_ROOT / "fluent-syntax" / "test"
+STRUCTURE_FIXTURES = FIXTURES_ROOT / "fixtures_structure"
+REFERENCE_FIXTURES = FIXTURES_ROOT / "fixtures_reference"
+
+ALL_FIXTURES = FluentCompatBase.collect_fixtures(
+  {json_dir: STRUCTURE_FIXTURES, ftl_dir: STRUCTURE_FIXTURES, category: :structure, with_spans: true},
+  {json_dir: REFERENCE_FIXTURES, ftl_dir: REFERENCE_FIXTURES, category: :reference, with_spans: false}
+)
 
 RSpec.describe "fluent-syntax Compatibility" do
   include FluentSyntaxCompatibility::TestHelper
 
-  FluentSyntaxCompatibility.all_fixtures.each do |fixture|
+  ALL_FIXTURES.each do |fixture|
     describe "#{fixture[:category]} fixtures" do
       context fixture[:name] do
         it "matches fluent-syntax AST output" do
