+++ spec/support/compat/bundle.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "bundle/ast_converter"
+require_relative "bundle/test_helper"
