+++ spec/support/compat/syntax.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "syntax/ast_comparator"
+require_relative "syntax/matchers"
+require_relative "syntax/test_helper"
