+++ spec/support/compat/syntax/ast_comparator.rb
