--- spec/support/fluent_bundle.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "fluent_bundle/ast_converter"
-require_relative "fluent_bundle/test_helper"
