--- spec/support/fluent_syntax.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "fluent_syntax/ast_comparator"
-require_relative "fluent_syntax/matchers"
-require_relative "fluent_syntax/test_helper"
